requirements.txt
attrs==21.4.0
bech32==1.2.0
black==22.3.0
cffi==1.15.0
click==8.1.2
iniconfig==1.1.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.1
pluggy==1.0.0
py==1.11.0
pycparser==2.21
PyNaCl==1.5.0
pyparsing==3.0.8
pytest==7.1.1
tomli==2.0.1
typing_extensions==4.1.1
